move without interview to its own component
