+++ frontend/src/Components/RecruitmentWithoutInterviewTable/components/WithoutInterview.module.scss
