+++ frontend/src/Components/RecruitmentWithoutInterviewTable/index.ts
@@ -1,0 +1,1 @@
+export { RecruitmentWithoutInterviewTable } from './RecruitmentWithoutInterviewTable';
