+++ frontend/src/Components/index.ts
@@ -40,6 +40,7 @@ export { ProtectedRoute } from './ProtectedRoute';
 export { PulseEffect } from './PulseEffect';
 export { RadioButton } from './RadioButton';
 export { RecruitmentApplicantsStatus } from './RecruitmentApplicantsStatus';
+export { RecruitmentWithoutInterviewTable } from './RecruitmentWithoutInterviewTable';
 export { SamfOutlet } from './SamfOutlet';
 export { SamfundetLogo } from './SamfundetLogo';
 export { SamfundetLogoSpinner } from './SamfundetLogoSpinner';
